+++ node_pool/CHANGELOG.md
@@ -1,3 +1,6 @@
+# node-pool-v3.3.1
+- Added `max_pods_per_node` paramater.
+
 # node-pool-v3.3.0
 - Added `node_metadata` parameter to control node metadata provided to workload, so that [workload identity](https://cloud.google.com/kubernetes-engine/docs/how-to/workload-identity) functionality may be used.